pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obd-toolkit"
version = "1.0.0"
description = "A comprehensive OBD2 diagnostic toolkit for vehicle diagnostics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "OBD Toolkit Developer"}
]
keywords = ["obd", "obd2", "diagnostics", "automotive", "elm327", "dtc", "vehicle"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    "obd>=0.7.1",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyserial>=3.5",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
obd-toolkit = "obd_toolkit.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/obd-toolkit"
Documentation = "https://github.com/yourusername/obd-toolkit#readme"
Repository = "https://github.com/yourusername/obd-toolkit"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
obd_toolkit = ["data/*.json"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=obd_toolkit --cov-report=term-missing"
